+++ src/main/java/simulator/DukeException.java
@@ -1,3 +1,5 @@
+package simulator;
+
 public class DukeException extends Exception{
 
     public DukeException(String msg) {
