+++ test/DeadlineTest.java
@@ -1,0 +1,10 @@
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class DeadlineTest {
+
+    @Test
+    void testToString() {
+    }
+}
